Apply bootstrap styles for storybook examples.
