+++ graylog2-web-interface/docs/graylog-luma/.storybook/withGraylogTheme.tsx
@@ -17,6 +17,7 @@
 import React from 'react';
 import 'preload';
 import 'theme/theme-styles';
+import 'bootstrap/less/bootstrap.less';
 import type { Decorator } from '@storybook/react';
 import { useDarkMode } from 'storybook-dark-mode';
 import { COLOR_SCHEME_DARK, COLOR_SCHEME_LIGHT } from '@graylog/sawmill';
